+++ tests/test_trade.py
@@ -24,6 +24,8 @@ class TradeTest(unittest.TestCase):
             self.mock_fill_kill,
             self.mock_offset,
             self.mock_green,
+            12,
+            34,
         )
 
     def test_init(self):
@@ -41,6 +43,8 @@ class TradeTest(unittest.TestCase):
         self.assertIsNotNone(self.trade.date_time_created)
         self.assertIsNone(self.trade.date_time_complete)
         self.assertIsNone(self.trade.market_notes)
+        self.assertEqual(self.trade.place_reset_seconds, 12)
+        self.assertEqual(self.trade.reset_seconds, 34)
 
     @mock.patch("flumine.order.trade.get_market_notes")
     def test_update_market_notes(self, mock_get_market_notes):
@@ -114,12 +118,22 @@ class TradeTest(unittest.TestCase):
     def test_create_order_from_current_limit_on_close(self):
         mock_current_order = mock.Mock()
         mock_current_order.order_type = "LIMIT_ON_CLOSE"
-        with self.assertRaises(NotImplementedError):
-            self.trade.create_order_from_current(mock_current_order, "12345")
+        order = self.trade.create_order_from_current(mock_current_order, "12345")
+        self.assertEqual(order.bet_id, mock_current_order.bet_id)
+        self.assertEqual(order.id, "12345")
+        self.assertEqual(self.trade.orders, [order])
 
     def test_create_order_from_current_market_on_close(self):
         mock_current_order = mock.Mock()
         mock_current_order.order_type = "MARKET_ON_CLOSE"
+        order = self.trade.create_order_from_current(mock_current_order, "12345")
+        self.assertEqual(order.bet_id, mock_current_order.bet_id)
+        self.assertEqual(order.id, "12345")
+        self.assertEqual(self.trade.orders, [order])
+
+    def test_create_order_from_current_unknown(self):
+        mock_current_order = mock.Mock()
+        mock_current_order.order_type = "BE"
         with self.assertRaises(NotImplementedError):
             self.trade.create_order_from_current(mock_current_order, "12345")